split `kernels/` into several files
